requirements.txt
aiogram==2.25.1
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.0
flask==3.0.0
requests==2.31.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
